Call the back end currency-exchange

+++ currency-exchange-service/src/main/java/com/meg/microservices/currencyexchangeservice/Repository/ExchangeRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ExchangeRepository extends JpaRepository<ExchangeValue, Integer> {
+public interface ExchangeRepository extends JpaRepository<ExchangeValue, Long> {
 
     ExchangeValue findByFromAndTo(String from, String to);
 }
